move alloc and gen support to tools
