+++ docs/conf.py
@@ -46,7 +46,8 @@ sys.path.append(os.path.abspath('../libensemble/alloc_funcs'))
 sys.path.append(os.path.abspath('../libensemble/gen_funcs'))
 sys.path.append(os.path.abspath('../libensemble/sim_funcs'))
 sys.path.append(os.path.abspath('../libensemble/comms'))
-sys.path.append(os.path.abspath('../libensemble/util'))
+sys.path.append(os.path.abspath('../libensemble/utils'))
+sys.path.append(os.path.abspath('../libensemble/tools'))
 sys.path.append(os.path.abspath('../libensemble/executors'))
 sys.path.append(os.path.abspath('../libensemble/resources'))
 # print(sys.path)
