+++ docs/examples/calling_scripts.rst
@@ -27,7 +27,7 @@ This example is from a test for evaluating the scaling capabilities of libEnsemb
 by calculating particle electrostatic forces through a user application. This
 application is registered with either the MPI or Balsam executor, then submitted
 for execution in the ``sim_f``. Note the use of the ``parse_args()`` and
-``save_libE_output()`` convenience functions from the :doc:`utilities<../utilities>`.
+``save_libE_output()`` convenience functions from the :doc:`tools<../utilities>` module.
 
 ..  literalinclude:: ../../libensemble/tests/scaling_tests/forces/run_libe_forces.py
     :language: python
