+++ libensemble/alloc_funcs/give_pregenerated_work.py
@@ -1,4 +1,4 @@
-from libensemble.alloc_funcs.support import avail_worker_ids, sim_work
+from libensemble.tools.alloc_support import avail_worker_ids, sim_work
 
 
 def give_pregenerated_sim_work(W, H, sim_specs, gen_specs, alloc_specs, persis_info):
