+++ libensemble/alloc_funcs/persistent_aposmm_alloc.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from libensemble.alloc_funcs.support import avail_worker_ids, sim_work, gen_work
+from libensemble.tools.alloc_support import avail_worker_ids, sim_work, gen_work
 
 
 def persistent_aposmm_alloc(W, H, sim_specs, gen_specs, alloc_specs, persis_info):
