+++ libensemble/alloc_funcs/start_fd_persistent.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from libensemble.alloc_funcs.support import avail_worker_ids, sim_work, gen_work, count_persis_gens
+from libensemble.tools.alloc_support import avail_worker_ids, sim_work, gen_work, count_persis_gens
 
 
 def finite_diff_alloc(W, H, sim_specs, gen_specs, alloc_specs, persis_info):
