+++ libensemble/tools/__init__.py
@@ -1,8 +1,7 @@
-# __all__ = ['libE_fields', 'save_libE_output', 'add_unique_random_streams',
-#            'check_inputs', 'ForkablePdb', 'parse_args']
-
 from .tools import save_libE_output, add_unique_random_streams, eprint
 from .check_inputs import check_inputs
 from .forkable_pdb import ForkablePdb
 from .parse_args import *
 from .fields_keys import *
+from .alloc_support import *
+from .gen_support import *
